requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
redis>=5.0.0
celery>=5.3.0
django-redis>=5.3.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
sentry-sdk>=1.32.0
geopy>=2.4.0
